refactor: browsecomp module
